pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tomatos"
version = "0.1"
requires-python = ">=3.9"
dependencies = [
    "pyyaml",
    "sklearn",
    "scikit-learn==1.1.1",
    "chex==0.1.3",
    "numpy==1.25.0",
    "jax==0.3.14",
    "scipy==1.9.2",
    "optax==0.1.2",
    "plothelp==0.1",
    "equinox==0.5.5",
    "pyhf @ git+https://github.com/fred-renner/pyhf.git@fix_jax",
    "relaxed @ git+https://github.com/fred-renner/relaxed.git@mle_fix",
    "neos==0.3.0",
    "jupyter",
    "h5py",
    "imageio",
    "imageio-ffmpeg",
    "uproot",
    "alive-progress",
    "vector",
    "isort",
]

[tool.setuptools.packages]
find = {} # Scan the project directory with the default parameters

[project.scripts]
tomatos = "tomatos.cli:main"
